[#30] refactor: 기능 기반 폴더 구조로 재배치 (features/, common/)
