+++ frontend/app/page.tsx
@@ -1,4 +1,4 @@
-import { SmartMirror } from "@/components/smart-mirror/smart-mirror";
+import { SmartMirror } from "@/features/smart-mirror";
 
 export default function Page() {
   return <SmartMirror />;
