+++ frontend/features/smart-mirror/components/audio-wave.tsx
