+++ frontend/features/smart-mirror/components/glass-panel.tsx
