+++ frontend/features/smart-mirror/components/p1-experience.tsx
