+++ frontend/features/smart-mirror/components/permission-guide.tsx
