+++ frontend/features/smart-mirror/components/scanning-effect.tsx
