+++ frontend/features/smart-mirror/components/webcam-background.tsx
