+++ frontend/features/smart-mirror/index.ts
@@ -1,0 +1,1 @@
+export { SmartMirror } from "./smart-mirror";
